feat: :recycle: refactor formateDateRange into helpers
