+++ packages/ripple-ui-core/src/index.ts
@@ -1,10 +1,9 @@
 export { default as rplEventBus } from './lib/eventbus'
 export { useAccessibleContainer } from './composables/useAccessibleContainer'
 export { useSlotContent } from './composables/useSlotContent'
-export { formatDateRange } from './lib/formatDateRange'
+export { formatDateRange, formatDate } from './lib/helpers'
 export { bpMin } from './lib/breakpoints'
 export { isExternalUrl } from './utils/isExternalUrl'
-export { formatDate } from './lib/helpers'
 export { RplIconGroups } from './components/icon/constants'
 export { useRippleEvent } from './composables/useRippleEvent'
 export { default as useEventContext } from './composables/useEventContext'
